pushing some results of refactoring. Now we call each of the DataContext flavors
