+++ great_expectations/data_context/data_context/ephemeral_data_context.py
@@ -1,6 +1,10 @@
+from abc import ABC
+from typing import Mapping, Optional, Union
+
 from great_expectations.data_context.data_context.abstract_data_context import (
     AbstractDataContext,
 )
+from great_expectations.data_context.types.base import DataContextConfig
 
 
 class EphemeralDataContext(AbstractDataContext):
@@ -12,4 +16,11 @@ class EphemeralDataContext(AbstractDataContext):
     compatibility reasons.
     """
 
-    pass
+    def __init__(
+        self,
+        project_config: Union[DataContextConfig, Mapping],
+        runtime_environment: Optional[dict] = None,
+    ):
+        super().__init__(
+            project_config=project_config, runtime_environment=runtime_environment
+        )
